requirements.txt
# Strands Agents with A2A support
strands-agents[a2a]
strands-agents-tools

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP client
httpx>=0.25.0

# YAML support
pyyaml>=6.0

# Diagrams for architecture visualization
diagrams>=0.23.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
